Issue #26
Link to issue: https://github.com/russelonato/ariel/issues/26

+++ Ariel/src/com/work/ariel/service/impl/BasicLaundrymat.java
@@ -80,11 +80,18 @@ public class BasicLaundrymat implements ILaundrymat {
 		commands.add(commandLines.get(0).replaceFirst(props.getProperty(CMD_SCRIPT, LPAR_NUMBER), param[0]));
 		commands.add(commandLines.get(1).replaceFirst(props.getProperty(CMD_SCRIPT, USERNAME), param[1]));
 		commands.add(commandLines.get(2).replaceFirst(props.getProperty(CMD_SCRIPT, PASSWORD), param[2]));
+
+		List<RPAInput> missingFiles = checkDownloadedFiles(param[4], rpaInputs);
+
+		if (missingFiles.size() == 0) {
+			return null;
+		}
+
 		if (param[3] != null && !param[3].trim().isEmpty()) {
 			commands.add(commandLines.get(3).replaceFirst(props.getProperty(CMD_SCRIPT, ENV_ASP_GROUP), param[3]));
 		}
 		commands.add(commandLines.get(4));
-		for (RPAInput rpaInput : checkDownloadedFiles(param[4], rpaInputs)) {
+		for (RPAInput rpaInput : missingFiles) {
 			String line = commandLines.get(5);
 			line = line.replaceFirst(props.getProperty(CMD_SCRIPT, LIBRARY), rpaInput.getLibrary());
 			line = line.replaceFirst(props.getProperty(CMD_SCRIPT, FILENAME), rpaInput.getFileName());
@@ -116,6 +123,10 @@ public class BasicLaundrymat implements ILaundrymat {
 
 		logger.logInfo("Generating Bat file");
 
+		if (file == null) {
+			return null;
+		}
+
 		try {
 			batFile = File.createTempFile(BAT_FILE, FileUtil.EXT_BAT);
 			batFile.deleteOnExit();
@@ -124,7 +135,8 @@ public class BasicLaundrymat implements ILaundrymat {
 		}
 
 		commands = new ArrayList<String>();
-		commands.add(commandLines.get(0).replaceFirst(props.getProperty(BAT, FILE), Matcher.quoteReplacement(file.getAbsolutePath())));
+		commands.add(commandLines.get(0).replaceFirst(props.getProperty(BAT, FILE),
+				Matcher.quoteReplacement(file.getAbsolutePath())));
 		commands.add(commandLines.get(1));
 
 		FileUtil.writeFile(batFile, commands);
@@ -163,7 +175,7 @@ public class BasicLaundrymat implements ILaundrymat {
 		rpaInputs = rpaInputManager.loadRPAInput(FileUtil.toFolder(param[4]));
 		launderer = new BasicLaunderer();
 		timeStart = new Date();
-		
+
 		logger.logInfo("*********Execution started*********");
 
 		Logger.getInstance().logInfo("Copying files.");
@@ -186,10 +198,15 @@ public class BasicLaundrymat implements ILaundrymat {
 			throw new SystemException(e.getMessage(), e.getCause());
 		}
 
-		Logger.getInstance().logInfo("Download files executed.");
-		BatExecutor batExecutor = new BatExecutor();
-		batExecutor
-				.execute(generateBat(generateCmdScript(rpaInputs, param[5], param[7], param[8], param[6], param[4])));
+		File batFile = generateBat(generateCmdScript(rpaInputs, param[5], param[7], param[8], param[6], param[4]));
+
+		if (batFile != null) {
+			Logger.getInstance().logInfo("Download files executed.");
+			BatExecutor batExecutor = new BatExecutor();
+			batExecutor.execute(batFile);
+		}else {
+			Logger.getInstance().logInfo("All files already exist. No downloads have been executed.");
+		}
 
 		Logger.getInstance().logInfo("Cleaning files.");
 		List<RPAInput> failedDownloads = checkDownloadedFiles(param[4], rpaInputs);
@@ -203,15 +220,16 @@ public class BasicLaundrymat implements ILaundrymat {
 						range);
 			}
 		}
-		
+
 		timeEnd = new Date();
-		
+
 		logger.logInfo("Execution completed. Total elapsed time: " + (timeEnd.getTime() - timeStart.getTime()) + " ms");
-		
-		if(failedDownloads.size() > 0) {
-			logger.logInfo("Execution has been completed but one or more files was not successfully downloaded. Please see list below: ");
-			
-			for(RPAInput failedDownload:failedDownloads) {
+
+		if (failedDownloads.size() > 0) {
+			logger.logInfo(
+					"Execution has been completed but one or more files was not successfully downloaded. Please see list below: ");
+
+			for (RPAInput failedDownload : failedDownloads) {
 				logger.logInfo(failedDownload.getMemberName());
 			}
 		}
